feat: :sparkles: Refactor weather app structure and add logging middleware
