+++ apps/weather/public/index.html
