+++ apps/weather/public/script.js
