+++ apps/weather/public/site.webmanifest
