+++ server.js
@@ -3,9 +3,12 @@ import cors from "cors";
 import helmet from "helmet";
 import bodyParser from "body-parser";
 
-import weatherAPI from "./apps/weather/routes.mjs";
+import weatherRoutes from "./apps/weather/routes/index.mjs";
+import logger from "./middleware/logger.mjs";
 
 const app = express();
+
+const NODE_ENV = process.env.NODE_ENV || "development";
 const PORT = process.env.PORT || 3000;
 const HOST = process.env.HOST || "localhost";
 const DB_URL = process.env.MONGO_URI;
@@ -16,15 +19,13 @@ app.use(cors());
 app.use(helmet());
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
-app.use(express.static("public"));
+app.use(logger);
 
-// Import Routes Here
-
-app.use("/weather/api", weatherAPI);
+app.use("/weather", weatherRoutes);
 
 app.listen(PORT, () => {
   console.group("\nStarting ZeroDae Apps Server...\n");
-  console.log(`NODE_ENV: ${process.env.NODE_ENV}`);
+  console.log(`NODE_ENV: ${NODE_ENV}`);
   console.log(`Hostname (Default: localhost): ${HOST}`);
   console.log(`Port (Default: 3000): ${PORT}`);
   console.log(`Database: ${DB_URL}`);
